Bug Fixes - Disposed Object
Bugs Problem should be a snapshot problem. Tween works over serverals seconds. In the meantime, it could happen from time to time, that the clients kills the enemey, while the server only have the old version of it. It seems that doubleblade was too fast or something so the problem arose there only now.